+++ CHANGES.md
@@ -61,6 +61,17 @@ To be released.
 [#110]: https://github.com/dahlia/fedify/issues/110
 
 
+Version 0.12.1
+--------------
+
+Released on July 27, 2024.
+
+ -  Fixed a bug where `fedify init -w hono` had generated scaffold files without
+    Fedify integration.
+ -  Fixed a bug where `fedify init -r bun -w hono` had generated scaffold files
+    with a wrong port number (was 3000).
+
+
 Version 0.12.0
 --------------